Cleanup & refactoring.
