+++ packages/cli/api-importers/openapi/openapi-v2-importer/src/3.1/paths/PathConverter.ts
@@ -4,7 +4,7 @@ import { TypeDeclaration } from "@fern-api/ir-sdk";
 
 import { AbstractConverter } from "../../AbstractConverter";
 import { ErrorCollector } from "../../ErrorCollector";
-import { methods } from "../../utils/consts";
+import { HttpMethods } from "../../constants/HttpMethods";
 import { OpenAPIConverterContext3_1 } from "../OpenAPIConverterContext3_1";
 import { OperationConverter } from "./OperationConverter";
 
@@ -42,7 +42,7 @@ export class PathConverter extends AbstractConverter<OpenAPIConverterContext3_1,
         // Check each HTTP method
         const inlinedTypes: Record<string, TypeDeclaration> = {};
 
-        for (const method of methods) {
+        for (const method of HttpMethods) {
             const operation = this.pathItem[method];
             if (operation != null) {
                 const operationConverter = new OperationConverter({
