+++ packages/cli/api-importers/openapi/openapi-v2-importer/src/constants/HttpMethods.ts
@@ -1,0 +1,1 @@
+export const HttpMethods = ["get", "post", "put", "delete", "patch"] as const;
