--- packages/cli/api-importers/openapi/openapi-v2-importer/src/utils/consts.ts
@@ -1,1 +1,0 @@
-export const methods = ["get", "post", "put", "delete", "patch"] as const;
